joins in subplan
